backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["eedu7 <mueedahmad067@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.112.0"}
uvicorn = {extras = ["standard"], version = "^0.30.5"}
pytest = "^8.3.2"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.32"
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
pydantic-extra-types = "^2.9.0"
phonenumbers = "^8.13.42"
python-dotenv = "^1.0.1"
alembic = "^1.13.2"
stubs = "^1.0.0"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
